move initializeIntegerSpinner to separate utils class
